requirements.txt
# Dependencies for the project  

chromadb
gensim==4.3.3
langchain==0.2.14
langchain_community
langchain_core==0.2.32
langchain_groq==0.1.9
langchain_huggingface==0.0.3
langchain_openai==0.1.21
plotly==5.23.0
python-dotenv==1.0.1
pymupdf
pypdf
streamlit==1.37.1
faiss-cpu

